fix hang when bug report is launched from flyout
