+++ src/runner/bug_report.h
@@ -1,0 +1,3 @@
+#pragma once
+
+void launch_bug_report() noexcept;
