+++ src/runner/runner.vcxproj.filters
@@ -42,6 +42,9 @@
     <ClCompile Include="centralized_hotkeys.cpp">
       <Filter>Utils</Filter>
     </ClCompile>
+    <ClCompile Include="bug_report.cpp">
+      <Filter>Utils</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="pch.h" />
@@ -85,6 +88,9 @@
       <Filter>Utils</Filter>
     </ClInclude>
     <ClInclude Include="centralized_hotkeys.h">
+      <Filter>Utils</Filter>
+    </ClInclude>
+    <ClInclude Include="bug_report.h">
       <Filter>Utils</Filter>
     </ClInclude>
   </ItemGroup>
